fix: Activityと Contentを分離
